Implement CGA layout rematerialization for Triton NVIDIA GPU dialect
- Introduced `cloneWithCGALayout` to clone blocked or sliced encodings with a new CGA layout while retaining access patterns.
- Added `getEncodingInCGA` to determine the appropriate encoding for values to maintain intra-CTA layout consistency.
- Created `CGARematerialization` class to handle the rematerialization of exclusively-used producer slices in the desired CTA layout.
- Updated `convertOpOperandsToLayouts` and layout assignment functions (`assignDotCTALayout`, `assignReduceCTALayout`) to support rematerialization roots.
- Modified tests to verify the absence of `ttg.force_cga_rematerialization` attributes in the output, ensuring correct rematerialization behavior.
