+++ lib/Dialect/TritonGPU/Transforms/RemoveLayoutConversions.cpp
@@ -24,7 +24,6 @@
 #include "triton/Dialect/TritonGPU/Transforms/Utility.h"
 #include <deque>
 #include <functional>
-#include <optional>
 
 namespace mlir::triton::gpu {
 
@@ -36,201 +35,6 @@ namespace mlir::triton::gpu {
 #define LDBG(X) LLVM_DEBUG(DBGS() << X << "\n")
 
 namespace {
-
-constexpr llvm::StringLiteral kForceCGARematerialization =
-    "ttg.force_cga_rematerialization";
-
-// Clone a blocked or sliced encoding with a new CGA layout while retaining the
-// layout choices that affect accesses within a CTA. The shape-aware blocked
-// layout builder redistributes threads and warps when the shape per CTA
-// changes.
-std::optional<Attribute> cloneWithCGALayout(RankedTensorType type,
-                                            CGAEncodingAttr cgaLayout,
-                                            Operation *scope) {
-  auto layout = dyn_cast<DistributedEncodingTrait>(type.getEncoding());
-  if (!layout)
-    return std::nullopt;
-
-  int numWarps = lookupNumWarps(scope);
-  OpBuilder builder(scope);
-  int threadsPerWarp = lookupThreadsPerWarp(builder);
-  if (auto blocked = dyn_cast<BlockedEncodingAttr>(layout)) {
-    return BlockedEncodingAttr::get(
-        type.getContext(), type.getShape(), blocked.getSizePerThread(),
-        blocked.getOrder(), numWarps, threadsPerWarp, cgaLayout);
-  }
-
-  if (auto slice = dyn_cast<SliceEncodingAttr>(layout)) {
-    auto parent = dyn_cast<BlockedEncodingAttr>(slice.getParent());
-    if (!parent)
-      return std::nullopt;
-    auto newParent = BlockedEncodingAttr::get(
-        type.getContext(), parent.getSizePerThread(),
-        parent.getThreadsPerWarp(), parent.getWarpsPerCTA(), parent.getOrder(),
-        cgaLayout);
-    return SliceEncodingAttr::get(type.getContext(), slice.getDim(), newParent);
-  }
-
-  return std::nullopt;
-}
-
-// Return the encoding that `value` should use to stay in the CTA group of
-// `preferredEncoding`. Dot operand encodings planned by AssignCTALayouts carry
-// additional parent-layout decisions, so reuse the preferred encoding when it
-// describes the same operand. Other producers retain their existing intra-CTA
-// layout and only adopt the preferred CGA layout.
-std::optional<Attribute>
-getEncodingInCGA(Value value, Attribute preferredEncoding, Operation *scope) {
-  auto type = dyn_cast<RankedTensorType>(value.getType());
-  if (!type)
-    return std::nullopt;
-
-  auto currentDot = dyn_cast<DotOperandEncodingAttr>(type.getEncoding());
-  auto preferredDot = dyn_cast<DotOperandEncodingAttr>(preferredEncoding);
-  if (currentDot && preferredDot &&
-      currentDot.getOpIdx() == preferredDot.getOpIdx() &&
-      currentDot.getKWidth() == preferredDot.getKWidth())
-    return preferredEncoding;
-
-  return cloneWithCGALayout(type, getCGALayout(preferredEncoding), scope);
-}
-
-// Rematerialize the producer slice of a marked conversion in the requested CTA
-// layout. Unlike the normal profitability-driven rematerializer, this rewrite
-// treats global loads as leaves: their tensor operands are converted to the new
-// blocked layout, and the load itself is cloned. This preserves coalescing
-// choices while avoiding a cross-CTA conversion of the loaded value.
-class ForcedCGARematerialization {
-public:
-  explicit ForcedCGARematerialization(ConvertLayoutOp root) : root(root) {}
-
-  LogicalResult run() {
-    auto desiredEncoding =
-        getEncodingInCGA(root.getSrc(), root.getType().getEncoding(), root);
-    if (!desiredEncoding || failed(plan(root.getSrc(), *desiredEncoding)) ||
-        !foundLoad)
-      return failure();
-
-    Value newSrc = rewrite(root.getSrc());
-    root.getSrcMutable().assign(newSrc);
-    return success();
-  }
-
-private:
-  LogicalResult plan(Value value, Attribute encoding) {
-    auto type = dyn_cast<RankedTensorType>(value.getType());
-    if (!type || type.getEncoding() == encoding)
-      return success();
-
-    auto [it, inserted] = layouts.try_emplace(value, encoding);
-    if (!inserted)
-      return success(it->second == encoding);
-
-    Operation *op = value.getDefiningOp();
-    if (!op || op->getNumResults() != 1 || !op->getRegions().empty())
-      return failure();
-
-    if (auto load = dyn_cast<LoadOp>(op)) {
-      // Duplicating a volatile load changes observable behavior when the
-      // original value has other users.
-      if (load.getIsVolatile())
-        return failure();
-      foundLoad = true;
-      return success();
-    }
-    if (isa<DescriptorLoadLikeOpInterface>(op)) {
-      foundLoad = true;
-      return success();
-    }
-
-    if (!isMemoryEffectFree(op) || !canBeRematerialized(op))
-      return failure();
-
-    if (auto gather = dyn_cast<GatherOp>(op))
-      return plan(gather.getIndices(), encoding);
-
-    if (auto convert = dyn_cast<ConvertLayoutOp>(op)) {
-      auto srcEncoding = getEncodingInCGA(convert.getSrc(), encoding, op);
-      if (!srcEncoding)
-        return failure();
-      return plan(convert.getSrc(), *srcEncoding);
-    }
-
-    for (auto [index, operand] : llvm::enumerate(op->getOpOperands())) {
-      if (!isa<RankedTensorType>(operand.get().getType()))
-        continue;
-      Attribute srcEncoding;
-      if (auto upcast = dyn_cast<UpcastFpOpInterface>(op))
-        srcEncoding = upcast.inferSrcEncoding(index, encoding);
-      else
-        srcEncoding = inferSrcEncoding(op, encoding);
-      if (!srcEncoding || failed(plan(operand.get(), srcEncoding)))
-        return failure();
-    }
-    return success();
-  }
-
-  Value rewrite(Value value) {
-    auto layoutIt = layouts.find(value);
-    if (layoutIt == layouts.end())
-      return value;
-    if (auto it = rewritten.find(value); it != rewritten.end())
-      return it->second;
-
-    Operation *op = value.getDefiningOp();
-    OpBuilder builder(op);
-    IRMapping mapping;
-
-    if (isa<LoadOp>(op)) {
-      for (OpOperand &operand : op->getOpOperands()) {
-        auto type = dyn_cast<RankedTensorType>(operand.get().getType());
-        if (!type)
-          continue;
-        auto newType = type.cloneWithEncoding(layoutIt->second);
-        Value converted = ConvertLayoutOp::create(builder, op->getLoc(),
-                                                  newType, operand.get());
-        mapping.map(operand.get(), converted);
-      }
-    } else if (!isa<DescriptorLoadLikeOpInterface>(op)) {
-      for (Value operand : op->getOperands()) {
-        if (isa<RankedTensorType>(operand.getType()))
-          mapping.map(operand, rewrite(operand));
-      }
-    }
-
-    Operation *newOp = builder.clone(*op, mapping);
-    newOp->removeAttr(kForceCGARematerialization);
-    for (auto [oldResult, newResult] :
-         llvm::zip(op->getResults(), newOp->getResults())) {
-      auto it = layouts.find(oldResult);
-      if (it == layouts.end())
-        continue;
-      auto oldType = cast<RankedTensorType>(oldResult.getType());
-      newResult.setType(oldType.cloneWithEncoding(it->second));
-      rewritten[oldResult] = newResult;
-    }
-    return rewritten.lookup(value);
-  }
-
-  ConvertLayoutOp root;
-  DenseMap<Value, Attribute> layouts;
-  DenseMap<Value, Value> rewritten;
-  bool foundLoad = false;
-};
-
-void forceCGARematerialization(ModuleOp module) {
-  SmallVector<ConvertLayoutOp> converts;
-  module.walk([&](ConvertLayoutOp convert) {
-    if (convert->hasAttr(kForceCGARematerialization))
-      converts.push_back(convert);
-  });
-
-  for (ConvertLayoutOp convert : converts) {
-    if (failed(ForcedCGARematerialization(convert).run()))
-      LDBG("failed forced CGA rematerialization for " << convert);
-    convert->removeAttr(kForceCGARematerialization);
-  }
-}
 
 // -----------------------------------------------------------------------------
 //
@@ -1809,10 +1613,6 @@ public:
   void runOnOperation() override {
     MLIRContext *context = &getContext();
     ModuleOp m = getOperation();
-
-    // Honor layout decisions made by AssignCTALayouts before canonicalization
-    // or forward propagation can rewrite the marked conversion boundaries.
-    forceCGARematerialization(m);
 
     // 1. Propagate layout forward starting from "anchor" ops.
     m.walk([](FuncOp funcOp) {
